feat: add quantity selector column to the table
This commit adds a new column to the table in the CheckoutB2B component. The new column displays the quantity selector information for each item in the cart.
