+++ react/CheckoutB2B.tsx
@@ -1,22 +1,24 @@
 import React, { useCallback } from 'react'
 import { useIntl } from 'react-intl'
-import { useCssHandles } from 'vtex.css-handles'
 import { OrderItems } from 'vtex.order-items'
+import { useCssHandles } from 'vtex.css-handles'
 import { ExtensionPoint, useRuntime } from 'vtex.render-runtime'
 import { Button, Layout, PageBlock, PageHeader, Table } from 'vtex.styleguide'
 
-import useOrderFormCustom from './hooks/useOrderFormCustom'
-import { useTableSchema, useTotalizers } from './utils'
-import { messages } from './utils/messages'
+import { messages } from './utils'
+import { useTableSchema, useOrderFormCustom, useTotalizers } from './hooks'
 
 function CheckoutB2B() {
   const handles = useCssHandles(['container'])
+
   const { loading, orderForm, setOrderForm } = useOrderFormCustom()
   const { items, totalizers, shipping, value: total, ...rest } = orderForm
+
   const { useOrderItems } = OrderItems
-  const { removeItem } = useOrderItems()
-  const mappedTotalizers = useTotalizers(totalizers, shipping, total)
   const schema = useTableSchema()
+  const { removeItem } = useOrderItems()
+
+  const mappedTotalizers = useTotalizers(totalizers, shipping, total)
 
   const { navigate } = useRuntime()
   const { formatMessage } = useIntl()
@@ -47,6 +49,7 @@ function CheckoutB2B() {
       >
         <PageBlock>
           <Table
+            fixFirstColumn
             totalizers={mappedTotalizers}
             loading={loading}
             fullWidth
@@ -56,6 +59,7 @@ function CheckoutB2B() {
             emptyStateLabel={formatMessage(messages.emptyCart)}
           />
         </PageBlock>
+
         {!!items.length && (
           <Button variation="danger-tertiary" onClick={handleClearCart}>
             {formatMessage(messages.clearCart)}
